functions lossless_jpeg_load_raw, ljpeg_start, getbithuff, fcol, smal_decode_segment, smal_v9_load_raw, redcine_load_raw, foveon_dp_load_raw, foveon_interpolate, crop_masked_pixels converted to dcraw 9.27

+++ Sources/ptDcRaw.h
@@ -477,7 +477,7 @@ public:
   uint16_t  m_IsFuji;
   int       m_Flip;
   int       m_Tiff_Flip;
-  int       m_Mask[8][4];
+  uint16_t  m_Mask[8][4];
   uint16_t  m_Curve[0x10000];
   uint16_t* m_Raw_Image;
   float     m_cmatrix[3][4];
